Reorganize files to separate custom reg fields from ui code
